+++ alacritty/src/main.rs
@@ -183,7 +183,14 @@ fn alacritty(mut options: Options) -> Result<(), Box<dyn Error>> {
     // Create the IPC socket listener.
     #[cfg(unix)]
     let socket_path = if config.ipc_socket() {
-        ipc::spawn_ipc_socket(&options, window_event_loop.create_proxy())
+        match ipc::spawn_ipc_socket(&options, window_event_loop.create_proxy()) {
+            Ok(path) => Some(path),
+            Err(err) if options.daemon => return Err(err.into()),
+            Err(err) => {
+                log::warn!("Unable to create socket: {:?}", err);
+                None
+            },
+        }
     } else {
         None
     };